Correções de interface
Foram corrigidas as interfaces da plataforma e das ONGs  a serem exibidas.

Foi criado um contexto para plataforma.

Foi corrigido a filtragem, agora permitindo escolher cidade e estado.
