+++ src/components/CitySelect/index.tsx
@@ -1,38 +1,33 @@
 import { useEffect, useState } from "react";
 import { Dropdown } from "react-bootstrap";
 
-import { useAuth } from "../../context/AuthContext";
+import { usePlatform } from "../../context/PlatformContext";
 
 export default function CitySelect() {
-  const { getCities, state, city, handleSetCity } = useAuth();
-  //const [selected, setSelected] = useState<string>(" - - ");
-  const [cities, setCities] = useState<string[]>([]);
+  const {setCity, getCities, setCities, cities, state, city} = usePlatform();
+
+  async function handleCities() {
+    const response = await getCities(state);
+    setCities(response)
+  }
 
   useEffect(() => {
-    try {
-      async function handleCities() {
-        console.log(state === ' - - ')
-        const response = await getCities( state === ' - - '  ? 'RJ' : state );
-
-        setCities(response.data);
+      if (state != 'Selecione um estado'){
+        handleCities();
+        setCity('Selecione uma cidade')
       }
-
-      handleCities();
-    } catch (err) {
-      console.log(err);
-    }
   }, [state]);
 
   return (
     <Dropdown
       onSelect={(event) => {
-      handleSetCity(event);
+      setCity(event);
       }}
     >
-      <Dropdown.Toggle style={{justifyContent:'center', minWidth:250}}>{` ${city} `}</Dropdown.Toggle>
+      <Dropdown.Toggle disabled={false} style={{justifyContent:'center', width:250}}>{` ${city} `}</Dropdown.Toggle>
 
       <Dropdown.Menu>
-        {cities.map((city) => (
+        {cities?.map((city) => (
           <Dropdown.Item eventKey={city}>{city}</Dropdown.Item>
         ))}
       </Dropdown.Menu>
