+++ src/components/Filters/index.tsx
@@ -1,4 +1,4 @@
-import {ListGroup, Card, Button} from 'react-bootstrap';
+import {ListGroup} from 'react-bootstrap';
 import CitySelect from '../CitySelect';
 import StateSelect from '../StateSelect';
