+++ src/components/Filters/styles.module.scss
@@ -1,7 +1,12 @@
-.container{
-    width: 100;
-    height: 95vh;
+.container {
+    padding-top: 3rem;
     display: flex;
+    width: 300px;
+    height: 93%;
+    margin-top: 3rem;
     flex-direction: column;
-}
-
+    background-color: #ffffff;
+    margin-right: 4rem;
+    justify-content: left;
+    border-radius: 0rem 2rem 0 0;
+  }
