+++ src/components/StateSelect/index.tsx
@@ -1,21 +1,19 @@
-import { useState } from "react";
 import { Dropdown } from "react-bootstrap";
-import { useAuth } from "../../context/AuthContext";
+import { usePlatform } from "../../context/PlatformContext";
 
 export default function StateSelect() {
-  const { state, handleSetState } = useAuth()
+  const {state, setState } = usePlatform()
   const json = require("./states.json");
 
   return (
     <Dropdown
-      onSelect={(event) => {
-        handleSetState(event);
+      onSelect={(event:string|undefined) => {
+        setState(event);
       }}
     >
-      <Dropdown.Toggle
-      style={{justifyContent:'center', minWidth:250}}
-      
-      >{` ${state} `}</Dropdown.Toggle>
+      <Dropdown.Toggle style={{justifyContent:'center', width:250}}>
+        {` ${state} `}
+      </Dropdown.Toggle>
 
       <Dropdown.Menu>
         {json.map((state) => {
