+++ src/context/AuthContext.tsx
@@ -11,22 +11,9 @@ type authContextData ={
     signUpWithApi:(informations:signUpTypes)=>Promise<void>;
     setToken:(token:string | undefined)=>void;
     getAccountInformation:()=>Promise<any>;
-    createDonation:(content:item)=>Promise<void>;
-    getInstitutions:(state:string, city?:string)=>Promise<any>;
-    getAllDonations:()=>Promise<any>;
-    getWaitDonation:(id:string)=>Promise<any>;
-    getCities:(state:string)=>Promise<any>;
     handleInstitutionsFiltered:({}) => void;
-    setUserPackage:(content:userPackageType)=>Promise<any>;
-    getAllUserPackage:()=>Promise<any>;
-    getUserPackage:(packageId:string)=>Promise<any>;
-    deleteUserPackage:(id:string)=>Promise<void>;
-    deleteContentUserPackage:(id:string)=>Promise<void>;
-    state:string;
-    handleSetState: (value: string) => void;
-    city: string;
-    handleSetCity: (value: string) => void;
 }
+
 
 export const AuthContext = createContext({} as authContextData);
 
@@ -49,31 +36,15 @@ type signUpTypes ={
     phone:string;
 }
 
-type item ={
-    name:string;
-    quantity:number;
-    priority:number;
-}
-
-type pkgitem ={
-    id:string; //id da doaçao
-    quantity:number;
-
-}
-
-type userPackageType ={
-    institution:string; //id da instituição
-    itens: pkgitem[];
-}
-
-
-
 export function AuthProvider({children}:authProviderProps){
     const [token, setToken] = useState<string | undefined>();
+    const [informations, setInformations] = useState<any[]>()
+
     const [userDonations, setUserDonations] = useState();
     const [institutionsFiltered, setInstitutionsFiltered] = useState()
-    const [state, setState] = useState<string>(' - - ')
-    const [city, setCity] = useState<string>(' - - ')
+
+    
+
     
     const router = useRouter();
 
@@ -94,6 +65,11 @@ export function AuthProvider({children}:authProviderProps){
     
                 setToken(data.data.token)
                 Cookies.set('token', token)
+                const info = await getAccountInformation()
+                Cookies.set('name', info.username)
+                console.log(info)
+                setInformations(info)
+                router.push('/platform')
 
         }catch(e){
             alert(`Email ou senha incorretos!`)
@@ -149,8 +125,7 @@ export function AuthProvider({children}:authProviderProps){
                         "Authorization":`${token}`
                     }
                 })
-
-                return data; // Dados do usuário completo{username, password, cpfCnpj/cpf ...}
+                return data.data; // Dados do usuário completo{username, password, cpfCnpj/cpf ...}
         }catch(e){
             console.log(e)
         }
@@ -158,192 +133,10 @@ export function AuthProvider({children}:authProviderProps){
 
   
 
-
-
-    //Cria uma doação
-    async function createDonation(content:item){
-        try{
-            await api_donations.post('',{
-                "description":content.name,
-                "quantity":content.quantity,
-                "priority":content.priority
-            },{
-                headers:{
-                    "Authorization":`${token}`
-                }
-            })
-
-        }catch{
-        }
-    }
-
-
-    async function getAllDonations(){ //Exibe todas doações necessárias para a empresa com CNPJ
-        try{
-            const data = await api_donations.get('/all', {
-                headers:{
-                    "Authorization":`${token}`
-                }
-            })
-            return data;
-        }catch(e){
-            console.log(e)
-        }
-    }
-
-    async function getInstitutions(state:string, city?:string) { //busca por 
-        try{
-            if(city){
-                const data = await api_institutions.get(``,
-                 {
-                    params:{
-                        "state":state,
-                        "city":city
-                    }}
-                 )
-
-                return data;
-            }
-            else{
-                const data = await api_institutions.get(``,
-                    {
-                        params:{
-                            "state":state
-                        }
-                    }
-                )
-                return data;
-            }
-        }catch(e){
-            console.log(e)
-        }
-    }
-
-
-    async function getWaitDonation(id:string){ //Exibe todas doação pendente de uma empresa
-        try{
-            const data = await api_donations.get(`/waiting_donator?id=${id}`, {
-            })
-            return data;
-
-        }catch(e){
-            console.log(e)
-        }
-    }
-
-    async function getCities(state:string) { 
-        
-        try{
-            const data = await api_cities.get('',
-                {
-                    params:{
-                        "state":state
-                    }
-                }
-            )
-
-            return data; // Dados do usuário completo{username, password, cpfCnpj/cpf ...}
-        }catch(e){
-            console.log(e)
-        }
-    }
-
-    async function setUserPackage(content:userPackageType) { //Coloca uma quantidade de um item e seu id no pacote.
-        console.log(content)
-        try{
-        await api_packages.post('',{
-                "institution":content.institution, 
-                "content": content.itens
-        },
-            {
-                headers:{
-                    "Authorization":`${token}`
-                }
-            }
-        )
-        }catch(e){
-            console.log(e)
-        }
-    }
-
-
-
-
-    //Retorna as doações que um usuario ja fez.
-    async function getAllUserPackage() { 
-        try{
-            const data = await api_packages.get('',
-                {
-                    headers:{
-                        "Authorization":`${token}`
-                    }
-                })
-            return data;
-        }catch(e){
-            console.log(e)
-        }
-    }
-
-
-    async function getUserPackage(packedId:string) { //retornaim pedido de um usuário
-        try{
-            const data = await api_packages.get('/content',
-                {
-                    params:{
-                        id:packedId
-                    },
-                    headers:{
-                        "Authorization":`${token}`
-                    }
-                })
-            return data;
-        }catch(e){
-            console.log(e)
-        }
-    }
-
-    async function deleteUserPackage(id:string){ //Deleta itens do package via id
-        try{
-             await api_packages.delete('', {
-                params:{
-                    "id":id
-                },
-                headers:{
-                    "Authorization":`${token}`
-                }
-            })
-        }catch(e){
-            console.log(e)
-        }
-    }
-
-    async function deleteContentUserPackage(id:string){ //Deleta itens co content da package via id
-        console.log(id)
-        try{
-             await api_packages.delete('/content', {
-                params:{
-                    "id":id
-                },
-                headers:{
-                    "Authorization":`${token}`
-                }
-            })
-        }catch(e){
-            console.log(e)
-        }
-    }
-
     function handleInstitutionsFiltered(value) {
         setInstitutionsFiltered(value)
     }
 
-    function handleSetState(value: string) {
-        setState(value)
-    }
-
-    function handleSetCity(value: string) {
-        setCity(value)
-    }
 
 
 
@@ -354,21 +147,7 @@ return(
         signInWithApi,
         signUpWithApi,
         getAccountInformation,
-        createDonation,
-        getInstitutions,
-        getAllDonations,
-        getWaitDonation,
         handleInstitutionsFiltered,
-        getCities,
-        setUserPackage,
-        getAllUserPackage,
-        getUserPackage,
-        deleteUserPackage,
-        deleteContentUserPackage,
-        state,
-        handleSetState,
-        city,
-        handleSetCity,
    
     }}>
         {children}
