+++ src/pages/platform/ong.module.scss
@@ -1,17 +1,25 @@
-.container {
+.ongContainer {
     display: flex;
-    justify-content: center;
+    height: 91.5vh;
+    width: 100%;
+    background-color: #2194c1;
+  
 
-    background-color: #2194C1;
 }
 
 .content {
-    width: 850px;
-    max-width: 100%;
-    justify-content: left;
+    width: 75.9%;
+    height: 93%;
+    display: flex;
+    justify-self: flex-end;
+    align-self: flex-end;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
     margin-top: 2.5rem;
+    overflow-y: scroll;
 
-    border-radius: 2rem 2rem 0 0;
+    border-radius: 2rem 0 0 0;
     background-color: #FFFFFF;
 
     h1 {
@@ -25,6 +33,7 @@
     margin: 2.5rem 0 0;
     border-radius: 1rem;
     background-color: #c4c4c4;
+    width: 50%;
 
     span {
         display: block;
@@ -33,11 +42,22 @@
 }
 
 .items {
+    display: flex;
     margin: 2.5rem 0 0;
     background-color: #c4c4c4;
+    width: 50%;
+    align-content: space-between;
 
-    h1 {
+    .ul{
+        flex:1
+        
+    }
+
+    .li{
         display: flex;
-        justify-content: center;
+        text{
+            flex: 1;
+        }
+        
     }
 }
